notification-service/requirements.txt
Flask==2.3.3
kafka-python==2.0.2
prometheus-client==0.17.1
python-dotenv==1.0.0